[ENH] Move shared rotation forest code to a common base class (#3630)
Stage 1 of the rotation forest speed up plan. Extracts the code duplicated
between RotationForestClassifier and RotationForestRegressor into a new
BaseRotationForest in aeon/base/_estimators/sklearn, following the
BaseIntervalForest pattern of switching task-specific behaviour on
is_classifier. The public classes are now thin subclasses keeping their
docstrings, signatures and public methods.

Outputs are bit-identical to the previous implementations for fit/predict/
predict_proba/fit_predict on both estimators, with one deliberate fix: the
regressor previously ignored its pca_solver parameter and always used the
'full' SVD solver. It now honours the parameter like the classifier, so the
expected values in test_rotf_output (which passes pca_solver='randomized')
have been regenerated.
