+++ aeon/base/_estimators/sklearn/__init__.py
@@ -1,0 +1,5 @@
+"""Base classes for sklearn-type estimators used in aeon."""
+
+__all__ = ["BaseRotationForest"]
+
+from aeon.base._estimators.sklearn._rotation_forest import BaseRotationForest
